+++ pyproject.toml
@@ -22,5 +22,5 @@ build-backend = "poetry.core.masonry.api"
 minversion = "6.0"
 addopts = "-rP -rx -s -vv"
 # testpaths = [
-#     "algofin/tests"
+#     "algofin/src/tests"
 # ]
